requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers==4.57.5
